fix: card wobbling when finger was moved off
